requirements.txt
# =============================================
# Fahasa AI Camera System - Requirements
# =============================================

# Core CV & AI
ultralytics>=8.3.0          # YOLOv11 (latest Ultralytics)
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# Face Analysis (Age / Gender)
deepface>=0.0.93
insightface>=0.7.3
onnxruntime>=1.17.0

# Multi-Object Tracking
lap>=0.4.0                  # ByteTrack dependency
filterpy>=1.4.5             # Kalman filter (DeepSORT)

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9      # PostgreSQL
redis>=5.0.0

# API / Backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# MLOps / Monitoring
mlflow>=2.11.0
prometheus-client>=0.20.0

# Analytics & Recommendation
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.12.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Messaging / Alerts
httpx>=0.27.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
schedule>=1.2.1
tqdm>=4.66.0
pyyaml>=6.0.1
